+++ src/engine/color.ts
@@ -1,4 +1,4 @@
-import { Random } from './math';
+import { lerp, Random } from './math';
 
 /**
  * Provides standard colors (e.g. {@apilink Color.Black})
@@ -48,6 +48,21 @@ export class Color {
     this.g = g;
     this.b = b;
     this.a = a != null ? a : 1;
+  }
+
+  public get hashCode(): number {
+    const r = Math.round(this.r) & 0xff;
+    const g = Math.round(this.g) & 0xff;
+    const b = Math.round(this.b) & 0xff;
+    const a = Math.round(this.a * 255) & 0xff;
+
+    let hash = 0;
+    hash = (hash << 5) - hash + r;
+    hash = (hash << 5) - hash + g;
+    hash = (hash << 5) - hash + b;
+    hash = (hash << 5) - hash + a;
+
+    return hash | 0;
   }
 
   /**
@@ -113,6 +128,17 @@ export class Color {
   public static fromHSL(h: number, s: number, l: number, a: number = 1.0): Color {
     const temp = new HSLColor(h, s, l, a);
     return temp.toRGBA();
+  }
+
+  /**
+   * Creates a new instance of Color from array of float components.
+   * Missing components will be replaced with 0 for r, g, b, and 1 for a.
+   * @param array  Array of [r, g, b, a] components
+   */
+  public static fromFloatArray(array: number[]): Color {
+    const components = [array[0] ?? 0, array[1] ?? 0, array[2] ?? 0].map((c) => Math.round(c * 255)) as [number, number, number];
+
+    return new Color(...components, array[3]);
   }
 
   /**
@@ -229,6 +255,25 @@ export class Color {
   }
 
   /**
+   * Return linear representation of a color component
+   * @param c color component
+   * @param scale color gamma, 2.2 recommended as standard
+   */
+  private static _COMPONENT_TO_LINEAR(c: number, scale: number = 2.2) {
+    return Math.pow(c, scale);
+  }
+
+  /**
+   * Return color component from its linear representation
+   * @param c color component
+   * @param scale color gamma, 2.2 recommended as standard
+   * @private
+   */
+  private static _COMPONENT_FROM_LINEAR(c: number, scale: number = 2.2) {
+    return Math.pow(c, 1.0 / scale);
+  }
+
+  /**
    * Return Hex representation of a color.
    */
   public toHex() {
@@ -258,6 +303,21 @@ export class Color {
   }
 
   /**
+   * Return float array representation of a color.
+   * @param precision
+   */
+  public toFloatArray(precision?: number): [r: number, g: number, b: number, a: number] {
+    let components = [this.r / 255, this.g / 255, this.b / 255];
+
+    if (precision) {
+      components = components.map((c) => parseFloat(c.toFixed(precision)));
+    }
+
+    components.push(this.a);
+    return components as [number, number, number, number];
+  }
+
+  /**
    * Returns a CSS string representation of a color.
    */
   public fillStyle() {
@@ -277,13 +337,61 @@ export class Color {
   }
 
   /**
-   * Lerp between two colors
+   * Lerp between two colors different modes:
+   * - hsl (default) - a compromise between speed and naturalness of the gradient, suitable for most cases;
+   * - rgb - the fastest algorithm, but worse results for complex gradients;
+   * - lrgb - the most realistic result, but slower than the others.
    */
-  public static lerp(colorA: Color, colorB: Color, t: number): Color {
+  public static lerp(colorA: Color, colorB: Color, t: number, colorSpace: 'hsl' | 'rgb' | 'lrgb' = 'hsl'): Color {
+    switch (colorSpace) {
+      case 'hsl':
+        return Color.lerpHSL(colorA, colorB, t);
+      case 'rgb':
+        return Color.lerpRGB(colorA, colorB, t);
+      case 'lrgb':
+        return Color.lerpLRGB(colorA, colorB, t);
+    }
+  }
+
+  /**
+   * Lerp between two colors using hsl as a compromise between speed and naturalness of the gradient
+   */
+  public static lerpHSL(colorA: Color, colorB: Color, t: number): Color {
     const color1: HSLColor = HSLColor.fromRGBA(colorA.r, colorA.g, colorA.b, colorA.a);
     const color2: HSLColor = HSLColor.fromRGBA(colorB.r, colorB.g, colorB.b, colorB.a);
     const newColor: HSLColor = HSLColor.lerp(color1, color2, t);
     return newColor.toRGBA();
+  }
+
+  /**
+   * Lerp between two colors using rgb for faster calculations
+   */
+  public static lerpRGB(colorA: Color, colorB: Color, t: number): Color {
+    return new Color(lerp(colorA.r, colorB.r, t), lerp(colorA.g, colorB.g, t), lerp(colorA.b, colorB.b, t), lerp(colorA.a, colorB.a, t));
+  }
+
+  /**
+   * Lerp between two colors using lrgb for more realistic gradient
+   */
+  public static lerpLRGB(colorA: Color, colorB: Color, t: number, gamma: number = 2.2): Color {
+    const rA = Color._COMPONENT_TO_LINEAR(colorA.r, gamma);
+    const gA = Color._COMPONENT_TO_LINEAR(colorA.g, gamma);
+    const bA = Color._COMPONENT_TO_LINEAR(colorA.b, gamma);
+
+    const rB = Color._COMPONENT_TO_LINEAR(colorB.r, gamma);
+    const gB = Color._COMPONENT_TO_LINEAR(colorB.g, gamma);
+    const bB = Color._COMPONENT_TO_LINEAR(colorB.b, gamma);
+
+    const rL = lerp(rA, rB, t);
+    const gL = lerp(gA, gB, t);
+    const bL = lerp(bA, bB, t);
+
+    return new Color(
+      Color._COMPONENT_FROM_LINEAR(rL, gamma),
+      Color._COMPONENT_FROM_LINEAR(gL, gamma),
+      Color._COMPONENT_FROM_LINEAR(bL, gamma),
+      lerp(colorA.a, colorB.a, t) // keeping alpha linear
+    );
   }
 
   public static random(rnd?: Random): Color {