+++ src/engine/index.ts
@@ -112,6 +112,7 @@ export * from './util/coroutine';
 export * from './util/assert';
 export * from './util/rental-pool';
 export * from './util/pause-system';
+export * from './util/serializer';
 
 // ex.Deprecated
 // import * as deprecated from './deprecated';